pyproject.toml
[project]
name = "evals-convex"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Convex", email = "sujayakar@convex.dev"},
]
dependencies = ["anthropic>=0.42.0", "python-dotenv>=1.0.1", "ipython>=8.31.0", "beautifulsoup4>=4.12.3", "requests>=2.32.3", "openai>=1.59.7", "markdown-it-py>=3.0.0", "portpicker>=1.6.0", "braintrust>=0.0.182", "autoevals>=0.0.117", "ruff>=0.9.3"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[dependency-groups]
dev = [
    "black>=24.10.0",
]
